Let inspect.isfunction() return True for functions.cache decorated functions
The initial motivation for this this change is to get DocTest's lineno properly
computed for functools.cache decorated function. This computation
happens in DocTest._find_lineno() which relies on inspect.isfunction()
to get function's code (and line numbers). Before this change,
_find_lineno() would return None for functools.cache decorated function.

We also need to unwrap() input function in inspect._has_code_flag() in
order to also get iscoroutinefunction() work for functools.cache
decorated functions.